Refactor institute_list model
- Moved functions related to `InstituteList` into the class
