+++ process_report/institute_list_models.py
@@ -1,9 +1,14 @@
 from typing import Annotated
 import datetime
 import functools
+import logging
 
 import pydantic
 import validators
+
+
+logger = logging.getLogger(__name__)
+logging.basicConfig(level=logging.INFO)
 
 
 def parse_date(v: str) -> str:
@@ -70,3 +75,27 @@ class InstituteList(pydantic.RootModel):
             if institute_info.courses_nonbillable:
                 institute_list.append(institute_info.display_name)
         return institute_list
+
+    @functools.cached_property
+    def domain_institute_mapping(self) -> dict[str, str]:
+        """Dict mapping web domains to institution display names"""
+        institute_map = dict()
+        for institute_info in self.root:
+            for domain in institute_info.domains:
+                institute_map[domain] = institute_info.display_name
+
+        return institute_map
+
+    def get_institution_from_pi(self, pi_email) -> str:
+        institution_domain = pi_email.split("@")[-1]
+        for i in range(institution_domain.count(".") + 1):
+            if institution_name := self.domain_institute_mapping.get(
+                institution_domain, ""
+            ):
+                break
+            institution_domain = institution_domain[institution_domain.find(".") + 1 :]
+
+        if institution_name == "":
+            logger.warning(f"PI name {pi_email} does not match any institution!")
+
+        return institution_name
